add validateEvmTokenApproval
