+++ packages/client/src/__test__/client.test.ts
@@ -2,14 +2,11 @@ import { rest } from "msw";
 import { setupServer } from "msw/node";
 
 import {
-  setApiOptions,
   chains,
   assets,
   assetsFromSource,
   bridges,
   BridgeType,
-  getFeeInfoForChain,
-  getRecommendedGasPrice,
   ibcOriginAssets,
   messages,
   recommendAssets,
@@ -26,12 +23,7 @@ import { toCamel } from "src/utils/convert";
 export const server = setupServer();
 
 describe("client", () => {
-  beforeEach(async () => {
-    await setApiOptions();
-  });
-  beforeAll(async () => {
-    server.listen();
-  });
+  beforeAll(async () => server.listen());
 
   afterEach(() => server.resetHandlers());
 
@@ -1530,26 +1522,29 @@ describe("client", () => {
     });
   });
 
-  describe("getRecommendedGasPrice", () => {
-    it("returns the recommended gas price for a chain", async () => {
-      const result = await getRecommendedGasPrice("osmosis-1");
+  // describe("getRecommendedGasPrice", () => {
+  //   it("returns the recommended gas price for a chain", async () => {
+  //     await setApiOptions();
+  //     const result = await getRecommendedGasPrice("osmosis-1");
 
-      expect(result?.denom).toEqual("uosmo");
-    }, 30000);
+  //     expect(result?.denom).toEqual("uosmo");
+  //   }, 30000);
 
-    it("returns the recommended gas price for Noble (no staking token)", async () => {
-      const result = await getRecommendedGasPrice("noble-1");
+  //   it("returns the recommended gas price for Noble (no staking token)", async () => {
+  //     await setApiOptions();
+  //     const result = await getRecommendedGasPrice("noble-1");
 
-      expect(result?.denom).toEqual("uusdc");
-    }, 30000);
-  });
+  //     expect(result?.denom).toEqual("uusdc");
+  //   }, 30000);
+  // });
 
-  describe("getFeeInfoForChain", () => {
-    it("returns the fee info for dymension", async () => {
-      const feeInfo = await getFeeInfoForChain("dymension_1100-1");
-      expect(feeInfo?.denom).toEqual("adym");
-    });
-  });
+  // describe("getFeeInfoForChain", () => {
+  //   it("returns the fee info for dymension", async () => {
+  //     await setApiOptions();
+  //     const feeInfo = await getFeeInfoForChain("dymension_1100-1");
+  //     expect(feeInfo?.denom).toEqual("adym");
+  //   });
+  // });
 
   describe("bridges", () => {
     it("returns a list of bridges", async () => {
