+++ packages/client/src/private-functions/cosmos/signCosmosMessageDirectInjective.ts
@@ -7,7 +7,7 @@ import { SignerData } from "@cosmjs/stargate/build/signingstargateclient";
 import { TxRaw } from "cosmjs-types/cosmos/tx/v1beta1/tx";
 import { CosmosMsg } from "src/types/swaggerTypes";
 import { getRestEndpointForChain } from "../getRestEndpointForChain";
-import { getEncodeObjectFromCosmosMessageInjective } from "../transactions";
+import { getEncodeObjectFromCosmosMessageInjective } from "./getEncodeObjectFromCosmosMessage";
 import { createTransaction } from "src/injective";
 import { fromBase64 } from "@cosmjs/encoding/build/base64";
 import { BigNumberInBase, DEFAULT_BLOCK_TIMEOUT_HEIGHT } from "@injectivelabs/utils";
